django/requirements.txt
asgiref==3.3.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-prometheus==2.1.0
django-yamlfield==1.1.0
djangorestframework==3.12.2
idna==3.2
oauthlib==3.1.1
prometheus-client==0.9.0
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.7